Move Import consumers to Core
