+++ src/Wallabag/CoreBundle/Consumer/AMQPEntryConsumer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Wallabag\ImportBundle\Consumer;
+namespace Wallabag\CoreBundle\Consumer;
 
 use OldSound\RabbitMqBundle\RabbitMq\ConsumerInterface;
 use PhpAmqpLib\Message\AMQPMessage;
