+++ src/Wallabag/CoreBundle/Consumer/RabbitMQConsumerTotalProxy.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Wallabag\ImportBundle\Consumer;
+namespace Wallabag\CoreBundle\Consumer;
 
 use OldSound\RabbitMqBundle\RabbitMq\Consumer;
